deleted bloat

+++ qsolvestuffs/validinput.h
@@ -1,4 +1,4 @@
 #include <stdio.h>
 #include <math.h>
 #include<stdio.h>
-int validDouble(double num);
+int validDouble(float num);
